[flow][records] Add constructor to record which takes in object with all fields
Summary:
Add constructor to record which takes in object with all fields. Any fields with a default value are optional.

This allows you to explicitly convert an object of the correct type into a record:
```
record R {
  a: number,
}

const o = {a: 1};

const x: R = new R(o);
```

E.g.
```
record R {
  a: number,
  b: string = '',
}
```
derives constructor
```
constructor({a: number, b?: string})
```

In the future, we will ban using `new` on a record with an object literal, e.g. `new R({a: 1})`, and this error will have a quickfix to use the record syntax `R {a: 1}`, so that is why the test cases do not directly use literals.

Error messages are not ideal yet (i.e. refering to "object type"), but it is a start for now.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D87098630

fbshipit-source-id: ed02c437b794ba0287b80c93fd0eff3301a544c9