[flow][records] Parsing for record expressions
Summary:
Adds parsing for record expressions. This is an identifier or member expression following by `{` (parsing as an object literal). E.g.

```
record R {
  a: number,
}

const x = R {a: 1};

const a = 2;
const y = R {a};

const z = R {...x};
```

The only place parsing record expressions is banned due to ambiguity is in the `extends` of a class - but this would be a Flow error anyway since:
- We already Flow error for any `extends` argument that is not an identifier or member expression, so would error for a record expression anyway
- Records cannot be extended, so it is nonsensical to put them in this position

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D87417107

fbshipit-source-id: 5d630f860cd03b2a62522ffe2a325a36104c29dc