[flow][records] Parse and type error on invalid semicolon suffix
Summary:
Parse and type error on invalid semicolon `;` suffix. This may come up as classes use semicolons as the seperator (while objects use comma `,` which we have chosen to use for records).

In the future, this will enable a quick-fix.

If there are multiple errors in one record delcaration, we combine the errors into one. This is written in a way to also allow additional invalid syntax errors in the future.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D87766642

fbshipit-source-id: a7b9d97cae86870d7caf8f565f96b1bc4029eeae