[flow][records] Store the set of defaulted props in the RecordKind
Summary:
Store the set of defaulted props (i.e. record properties with a default value, e.g. `foo: number = 0`) in the `RecordKind`, so we are able to access this information from the type.
While not used when accessing properties, this is useful for setting the constructor and also autocomplete.

Changelog: [internal]

Reviewed By: marcoww6

Differential Revision: D87901066

fbshipit-source-id: f3dce97b116c5ff293925dd3d89070503a7071f3