[flow][records] Update ESTree output of record declaration sub-nodes
Summary:
Update ESTree output of record declaration sub-nodes. We have both `RecordDeclaration` and `RecordExpression`,
so the sub-nodes of `RecordDeclaration` should be clear that they are related to the declaration and not the expression.

Also, rename `RecordDeclarationBody` property `body` to `elements` to better capture the output, and so we don't have `body.body`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D88059452

fbshipit-source-id: 2bb06dfdee1cca27dde7000290a582f3f7af6186