[flow][upgradeCodemod] Bug Fix
Summary:
For whatever reason this codemod only works in one file and all subsequent files will have the same output as the output of the first file.

Claude code came up with this solution and it seems reasonable to me. Now `jest test` only have 1 test failed and it was clown town previously

Reviewed By: SamChou19815

Differential Revision: D88983452

fbshipit-source-id: 3569e720596d56d01ab2fb8e5d3470a9ffa1fc6a