[flow] Support well known symbols in class methods
Summary:
This diff supports well-known symbols in class methods. For now, we will require you to use the following exact syntax:

```
class MyIterable {
  *[Symbol.iterator](): Iterator<number> {
    yield 1;
    yield 2;
    yield 3;
  }
}
```

instead of allowing anything that has the same type of `Symbol.iterator` to fill the slot of the `Symbol.iterator` expression. It will be a todo for the future.

For now, I go with the simple approach of recognizes these pattern syntactically, and translate them into the hacky old `@properly` identifier name. As a result, the error message is still a bit rough.

Reviewed By: panagosg7

Differential Revision: D89075707

fbshipit-source-id: 42a15dc45408f2a4ca9ae09f17113c7e11b776d1