[flow] Support omitting `declare` for `declare const/let/var` under declaration context
Summary:
This is part of the effort to help flow consume .d.ts directly. Currently, the biggest issue to consume .d.ts as is is that flow requires declare for all the declarations, even if it's clear that things are in the declaration context.

This diff starts to change that, we start with `declare const/let/var`. This is relatively easy, since it requires no AST change, AST can already represent `const/let/var foo: string` (although for `const foo: string`, it's a parse error due to validation).

All we need to do is to setup the infra to tell whether we are in declaration context, and then go through the normal parsing of variable declaration. In the end, we only need to make a decision after we parsed it. If we are in declaration context, then we allow a const without binding expression.

Changelog: [feature] `declare` keyword can be omitted for `declare const/let/var variable: type` under declaration context (libdef file, .js.flow file, declare module, declare namespace)

Reviewed By: gkz

Differential Revision: D89518680

fbshipit-source-id: a23e35c38833e185bf3e359c5aed693abbdcaf3d